Move the cuda and dnn flags in configdefaults since the later depends on the former.
Otherwise it is possible to get into a situation where we use the dnn flags, but the cuda ones aren't defined and stuff breaks.
